Update OmniWheelsDevon.java
assigned actuators to buttons and directions. Set all servos to auto go to 0.0 so it does "something".

Next need is to figure out how I want to code the arms and actuators, and start coding it. (Likely-hood is top arm is up and down on a joystick, might need to move code around, actuators are triggers, and bottom arm is just a and b)